fix majority of failed LLMs (#1247)
* fix majority of failed LLMs

* fix awq and glm4

* Update optimum/exporters/openvino/convert.py

+++ tests/openvino/utils_tests.py
@@ -281,9 +281,8 @@ def patch_awq_for_inference(to_patch):
 
             out_shape = x.shape[:-1] + (out_features,)
             x = x.to(torch.float16)
-
-            out = dequantize_gemm(qweight, qzeros, scales, w_bit, group_size)
-            out = torch.matmul(x, out)
+            out = dequantize_gemm(qweight.to(torch.int32), qzeros.to(torch.int32), scales, w_bit, group_size)
+            out = torch.matmul(x, out.to(x.dtype))
 
             out = out + bias if bias is not None else out
             out = out.reshape(out_shape)
